% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{BalanceData}
\alias{BalanceData}
\title{Balance Data}
\usage{
BalanceData(Raw_data, Label, Type = "ubUnder")
}
\arguments{
\item{Raw_data}{Raw data but no label}

\item{Label}{Label of the data}

\item{Tyep}{Choose what method to balance the data}
}
\value{
Balanced data
}
\description{
Balance Data
}
