% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{Caret_Model}
\alias{Caret_Model}
\title{Train A Model}
\usage{
Caret_Model(train, test, model, search = F)
}
\arguments{
\item{train}{train data}

\item{test}{test data}

\item{model}{which model you chioce}

\item{search}{random search}
}
\value{
no return
}
\description{
Train A Model
}
