% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{IV_WOE}
\alias{IV_WOE}
\title{Calculate IV and perform woe transformation}
\usage{
IV_WOE(Raw_data, Label, Woe_t = F, Filter = 0)
}
\arguments{
\item{Raw_data}{Raw data but no label}

\item{Label}{Label of the data}

\item{Woe_t}{The default is F, indicating whether to return the data after the woe transformation}

\item{Filter}{a value indicating that the variable whose IV is greater than the value is filtered out}
}
\value{
Returns the IV value of an l variable, the variable transformed by woe, the name of the feature after filtering...
}
\description{
Calculate IV and perform woe transformation
}
