% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{Mlr_Modle}
\alias{Mlr_Modle}
\title{Train A Model}
\usage{
Mlr_Modle(Train, Test, model)
}
\arguments{
\item{model}{which model you chioce}

\item{train}{train data}

\item{test}{test data}
}
\description{
Train A Model
}
