% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{SplitSample}
\alias{SplitSample}
\title{Dividing data sets}
\usage{
SplitSample(Raw_data, Label, rate)
}
\arguments{
\item{Raw_data}{Raw data}

\item{Label}{label}

\item{rate}{the rate to split data}
}
\value{
return train and test
}
\description{
Dividing data sets
}
